Main Page: Refactored login to seperate smaller chuncks and reworked the login flow. Was looping through access code refreshes without actually rechecking the access code each time.
Now tries the access code, if it fails, gets a new one, then resets the acesscode before trying again.

Timer:
Changed all of the ints to doubles to match the TimeSpan settings, updated some of the model to be TimeSpans instead of ints, and added Elapsed Time and Total Time counters.