Quality of Life Added a Messages section to the Home page to easily test connection and updated the username at the top so it doesnt display a full email address.
Major
Added home and Timer to a Group when they join SignalR Hubs based on the Users ID. This way all hub instances that User joins will be linked to them, and they will only recieve messages for their own hubs.

TODO Only set for messages so will have to apply this to all events.