Made some bigger changes.
Renamed twitchHub => streamHub as I want to use one main hub for most of the major actions related to running the app. May divide it more later, but want to avoid 2-3 hubs per widget.

StreamTimer is now working. Has a single background timer for the whoe system, and all widgets will tick on the same tick. Current Timer widget catches a tick, calls tick in the background class, and ticks the clock.

Background class will handle most of the information, and front end component will just display the information. Add the component to a page and call the background tasks to actually run the clock. Means just the clock can be shown and controls can be somewhere else.